rename new one to statsreact, some staging
