--- statsreact.new/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "statsreact",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@types/jest": "24.0.18",
-    "@types/node": "12.7.4",
-    "@types/react": "16.9.2",
-    "@types/react-dom": "16.9.0",
-    "create-react-class": "^15.6.0",
-    "eslint": "^6.0.0",
-    "jquery": "^3.2.1",
-    "jquery-ui": "^1.12.1",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-scripts": "3.1.1",
-    "react-slider": "^0.11.2",
-    "react-transition-group": "^1.1.3",
-    "typescript": "3.6.2",
-    "wicked-good-xpath": "^1.3.0"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
-}
