+++ statsreact/package.json
@@ -1,29 +1,44 @@
 {
-  "name": "statsnew",
-  "version": "1.0.0",
-  "description": "Describe statsnew here",
+  "name": "statsreact",
+  "version": "0.1.0",
   "private": true,
-  "scripts": {
-    "build": "nwb build-react-app && chmod +x dist/stats.html",
-    "clean": "nwb clean-app",
-    "start": "nwb serve-react-app",
-    "test": "nwb test-react",
-    "test:coverage": "nwb test-react --coverage",
-    "test:watch": "nwb test-react --server"
-  },
+  "homepage": "https://gregstoll.dyndns.org/~gregstoll/baseball.test",
   "dependencies": {
+    "@types/jest": "24.0.18",
+    "@types/node": "12.7.4",
+    "@types/react": "16.9.2",
+    "@types/react-dom": "16.9.0",
+    "create-react-class": "^15.6.0",
+    "eslint": "^6.0.0",
     "jquery": "^3.2.1",
     "jquery-ui": "^1.12.1",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-slider": "^0.8.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
+    "react-scripts": "3.1.1",
+    "react-slider": "^0.11.2",
     "react-transition-group": "^1.1.3",
+    "typescript": "3.6.2",
     "wicked-good-xpath": "^1.3.0"
   },
-  "devDependencies": {
-    "nwb": "0.16.x"
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build && node replaceSsiTags.js",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
-  "author": "",
-  "license": "MIT",
-  "repository": ""
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
