+++ statsreact/public/index.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8" /> 
+    <meta name="msApplication-ID" content="57039gregstoll.com.BaseballCalculator"/>
+    <meta name="msApplication-PackageFamilyName" content="57039gregstoll.com.BaseballCalculator_mrfc7nnpdd326"/>
+    <link rel="canonical" href="https://gregstoll.com/~gregstoll/baseball/stats.html" />
+    <title>Win Expectancy Finder</title>
+    <meta name="viewport" content="width=device-width">
+    <ssihead />
+</head>
+<body>
+  <ssibodytop />
+  <div class="container">
+      <div id="root"></div>
+<p style="clear: both;">
+  <a href='https://play.google.com/store/apps/details?id=com.gregstoll.BaseballStats&pcampaignid=MKT-Other-global-all-co-prtnr-py-PartBadge-Mar2515-1'><img alt='Get it on Google Play' src='https://play.google.com/intl/en_us/badges/images/generic/en_badge_web_generic.png' style="width: 240px; height: 93px;"/></a>
+<a href="https://www.microsoft.com/store/apps/9nblgggzkgt4"><img src="https://cmsresources.windowsphone.com/devcenter/en-us/legacy_v1/img/badgegenerator/English_wstore_black_258x67.png" alt="Download from Windows Store" /></a> <a href="http://www.windowsphone.com/s?appid=ab476634-9a24-4f0c-bc7a-4f65a12c33dd"><img src="DownloadWPMed.png" alt="Download from the Windows Phone Store" title="Download from the Windows Phone Store"></a></p>
+<p>To make a graph of the Win Expectancy of a baseball game, use the <a href="graphgame.html">graph a game</a> tool.</p>
+<p>This data is from MLB games from 1957-2018, including postseason games.  All data is available at <a href="http://www.retrosheet.org/game.htm">Retrosheet</a>.</p>
+<p><b>Linked from:</b> <a href="https://fivethirtyeight.com/features/donald-trump-has-a-20-percent-chance-of-becoming-president/">FiveThirtyEight</a> (search for "eighth inning"), <a href="https://www.washingtonpost.com/news/dc-sports-bog/wp/2016/05/31/a-seven-run-comeback-a-27-hour-rain-delay-and-a-trip-to-the-ncaa-tournament/">The Washington Post</a>, <a href="http://www.cbssports.com/mlb/news/how-win-expectancy-lets-you-track-each-teams-chances-to-win-world-series-game-7/">CBS Sports</a>, <a href="http://www.cbssports.com/mlb/eye-on-baseball/24732478/which-of-the-many-sac-bunts-in-al-wild-card-game-made-sense">CBS Sports again</a>, <a href="http://www.royalsreview.com/2016/4/4/11365240/fun-with-wp-and-wpa">Royals Review</a>, <a href="http://www.beyondtheboxscore.com/2015/10/21/9574081/the-impact-of-managerial-mistakes">Beyond the Box Score</a>, and more!</p>
+<p>Idea taken from walkoffbalk.com's Win Expectancy Finder (now defunct).  Here's a <a href="http://www.hardballtimes.com/main/article/the-one-about-win-probability/">good article about Win Expectancy</a>.  Leverage Index calculations based on <a href="http://www.hardballtimes.com/crucial-situations/">this article</a>.</p>
+<p>Source files available on <a href="https://github.com/gregstoll/baseballstats">GitHub</a>.</p>
+<p>Articles written with this data:</p>
+<ul>
+    <li><a href="https://gregstoll.wordpress.com/2019/05/29/how-common-are-walk-off-walks-on-four-pitches-in-baseball/">How common are walk-off walks (on four pitches!) in baseball?</a></li>
+</ul>
+<p>The information used here was obtained free of charge from and is copyrighted by Retrosheet.  Interested parties may contact Retrosheet at 20 Sunset Rd., Newark, DE 19711.</p>
+<p><a href="index.html">Baseball statistic calculators</a></p>
+<script src="https://www.google-analytics.com/urchin.js" type="text/javascript"></script>
+<ssibodybottom />
+</div>
+</body></html>
