+++ statsreact/public/manifest.json
