+++ statsreact/src/App.css
@@ -1,58 +1,33 @@
 .App {
-  height: 100%;
-  min-height: 400px;
   text-align: center;
+}
+
+.App-logo {
+  animation: App-logo-spin infinite 20s linear;
+  height: 40vmin;
+  pointer-events: none;
+}
+
+.App-header {
+  background-color: #282c34;
+  min-height: 100vh;
   display: flex;
   flex-direction: column;
-  justify-content: stretch;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
 }
 
-P.littlespace {margin: 2px;}
-.leverageVeryHigh {background-color: #f22; font-weight: 900;}
-.leverageHigh {background-color: #99f; font-weight: 900;}
-.leverageMedium {background-color: #ccc; font-weight: 900;}
-table.scoreTable td { /*border: 1px solid black;*/ font-size: 150%;}
-table.scoreTable td { padding: 1px 5px 1px 5px; }
-
-table.innings { width: 300px; }
-table.innings td { padding: 1px; width: 9%; text-align: center; }
-table.innings td.sideHeader { width: auto; }
-table.innings input[type="radio"] {
-   border-radius: 100% !important;
-   margin-block-end: 0;
-   margin-block-start: 3px;
-   margin-inline-end: 3px;
-   margin-inline-start: 5px;
-}
-table.innings th {
-   text-align: center;
+.App-link {
+  color: #61dafb;
 }
 
-table.runsPerInning td,th {
-   text-align: right;
-   padding: 0px 3px 0px 3px;
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
 }
-
-/* transitions */
-.outputTransition-enter {
-  opacity: 0.01;
-}
-.outputTransition-enter.outputTransition-enter-active {
-  opacity: 1;
-  transition: opacity 0.5s ease-in;
-}
-.outputTransition-leave {
-  opacity: 1;
-}
-.outputTransition-leave.outputTransition-leave-active {
-  opacity: 0.01;
-  transition: opacity 0.2s ease-in;
-}
-/*.outputTransition-appear {
-      opacity: 0.01;
-      transition: opacity .5s ease-in;
-}
-.outputTransition-appear.outputTransition-appear-active {
-      opacity: 1;
-}*/
-
