+++ statsreact/src/Appstock.tsx
