+++ statsreact/src/serviceWorker.ts
