+++ statsreact/tsconfig.json
